// archive/legacy-review/idefix-backoffice/Idefix Backoffice/gstech-backoffice/src/js/core/api/types.ts
export type FetchApi = (
  url: string,
  config?: RequestInit & {
    params?: { [key: string]: any } | null;
  },
) => Promise<any>;

export interface AuthenticationAPI {
  login: (data: LoginCredentials) => Promise<LoginResponse>;
  logout: () => Promise<{}>;
  expirePassword: () => Promise<{}>;
}

export interface CampaignsAPI {
  getPlayerCampaigns: (values: NewCampaign) => Promise<Player[]>;
  getAffiliates: () => Promise<Affiliate[]>;
}

export interface KycAPI {
  get: (playerId: number) => Promise<Kyc[]>;
  getDocument: (playerId: number, kycDocumentId: number) => Promise<Kyc>;
  create: (playerId: number, photos: { id: string; originalName: string }[]) => Promise<KycDocument[]>;
  createContent: (playerId: number, content: string) => Promise<KycDocument[]>;
  verify: (playerId: number, kycDocumentId: number, document: DocumentBase) => Promise<OKResponse>;
  update: (playerId: number, kycDocumentId: number, documentDraft: DocumentDraft) => Promise<KycDocument>;
  decline: (playerId: number, kycDocumentId: number) => Promise<OKResponse>;
  updatePhoto: (playerId: number, kycDocumentId: number, documentDraft: DocumentDraft) => Promise<KycDocument>;
  requestDocuments: (playerId: number, values: KycRequestDraft) => Promise<PlayerAccountStatus>;
  // getKycRequest: (playerId: number) => Promise<KycRequest[]>;
}